# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Innovo Labs Digital Initiatives Portal API
  version: "1.0.0"
  description: >
    REST API for the Innovo Labs portal. All endpoints are Next.js Route Handlers
    under `/api/**`, running as serverless functions (Node.js runtime) on Vercel.

    Authentication: session is a JWT (HS256, jose) stored in an HttpOnly cookie
    named `innovo_session`, issued after Microsoft Entra ID (OAuth2/OIDC) SSO.
    Because the cookie is HttpOnly it is sent automatically by the browser; there
    is no bearer token in the Authorization header.

    Authorization: enforced in two layers — the `defineRoute` wrapper (role gate +
    Zod validation + rate limit) and PostgreSQL Row-Level Security.

    Common responses (returned by the shared wrapper):
      401 Not authenticated · 403 Forbidden · 400 validation error ·
      429 Too Many Requests · 500 Internal server error.
servers:
  - url: https://<your-vercel-domain>
    description: Production (Vercel)
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: innovo_session
      description: HttpOnly JWT session cookie issued after Entra ID SSO.
  responses:
    Unauthorized:
      description: No valid session cookie.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authenticated but role/permission gate failed.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
security:
  - sessionCookie: []
paths:
...
  /api/auth/callback/azure:
    get:
      tags: [Auth]
      security: []
      summary: SSO callback
      description: Validates `state` against the cookie, exchanges the code, issues the session cookie.
      parameters:
        - { in: query, name: code, schema: { type: string }, required: true }
        - { in: query, name: state, schema: { type: string }, required: true }
      responses:
        "302": { description: Redirect to the app (or /login with an error param). }
...
  /api/startups:
    get:
      tags: [Startups]
      summary: List startups (paginated)
      parameters:
        - { in: query, name: bu, schema: { type: string } }
        - { in: query, name: dept, schema: { type: string } }
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        "200": { description: OK. }
    post:
      tags: [Startups]
      summary: Create a startup (innovation_admin / super_admin)
      responses:
        "201": { description: Created. }
        "403": { $ref: '#/components/responses/Forbidden' }
        "409": { description: Duplicate company name. }
...
  /api/challenges:
    get: { tags: [Content], summary: List challenges, responses: { "200": { description: OK } } }
    post: { tags: [Content], summary: Create challenge (admin), responses: { "201": { description: Created }, "403": { $ref: '#/components/responses/Forbidden' } } }
...
  /api/news/upload-video:
    post:
      tags: [Content]
      summary: Upload a news video (authenticated)
      requestBody: { content: { multipart/form-data: { schema: { type: object, properties: { file: { type: string, format: binary } } } } } }
      responses:
        "200": { description: Uploaded (returns URL). }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /api/upload:
    post:
      tags: [Files]
      summary: Upload a file/image (authenticated) → returns a storage URL
      requestBody: { content: { multipart/form-data: { schema: { type: object, properties: { file: { type: string, format: binary } } } } } }
      responses:
        "200": { description: Uploaded. }
        "401": { $ref: '#/components/responses/Unauthorized' }